Improve the application layout
