+++ data/ui/widgets.cmb
@@ -2,7 +2,6 @@
 <!DOCTYPE cambalache-project SYSTEM "cambalache-project.dtd">
 <cambalache-project version="0.13.1" target_tk="gtk-4.0">
   <ui>
-	(1,None,"widgets.ui","widgets.ui",None,None,None,None,None,None,None),
 	(2,1,"new_page.ui","new_page.ui",None,None,None,None,None,None,None),
 	(3,1,"main.ui","main.ui",None,None,None,None,None,None,None),
 	(4,4,"locator.ui","locator.ui",None,None,None,None,None,None,None),
@@ -74,7 +73,8 @@
 	(6,8,"GtkButton","next_btn",5,None,None,None,2,None),
 	(6,9,"GtkLabel","search_count_label",5,None,None,None,3,None),
 	(6,10,"WebKitWebView","web_view",3,None,None,None,1,None),
-	(6,11,"WebKitSettings",None,10,None,None,None,None,None)
+	(6,11,"WebKitSettings",None,10,None,None,None,None,None),
+	(6,12,"GtkBox","sidebar",2,None,None,None,None,None)
   </object>
   <object_property>
 	(2,4,"GtkBox","spacing","20",None,None,None,None,None,None,None,None,None),
@@ -143,6 +143,7 @@
 	(5,6,"GtkLabel","ellipsize","start",None,None,None,None,None,None,None,None,None),
 	(5,6,"GtkWidget","margin-bottom","20",None,None,None,None,None,None,None,None,None),
 	(6,2,"AdwOverlaySplitView","content",None,None,None,None,None,3,None,None,None,None),
+	(6,2,"AdwOverlaySplitView","sidebar",None,None,None,None,None,12,None,None,None,None),
 	(6,3,"GtkOrientable","orientation","vertical",None,None,None,None,None,None,None,None,None),
 	(6,3,"GtkWidget","hexpand","True",None,None,None,None,None,None,None,None,None),
 	(6,3,"GtkWidget","hexpand-set","True",None,None,None,None,None,None,None,None,None),
@@ -163,7 +164,10 @@
 	(6,11,"WebKitSettings","enable-back-forward-navigation-gestures","True",None,None,None,None,None,None,None,None,None),
 	(6,11,"WebKitSettings","enable-media-capabilities","True",None,None,None,None,None,None,None,None,None),
 	(6,11,"WebKitSettings","enable-media-stream","True",None,None,None,None,None,None,None,None,None),
-	(6,11,"WebKitSettings","enable-spatial-navigation","True",None,None,None,None,None,None,None,None,None)
+	(6,11,"WebKitSettings","enable-spatial-navigation","True",None,None,None,None,None,None,None,None,None),
+	(6,12,"GtkOrientable","orientation","vertical",None,None,None,None,None,None,None,None,None),
+	(6,12,"GtkWidget","vexpand","True",None,None,None,None,None,None,None,None,None),
+	(6,12,"GtkWidget","vexpand-set","True",None,None,None,None,None,None,None,None,None)
   </object_property>
   <object_signal>
 	(1,3,12,"GtkButton","clicked","new_tab",None,None,None,None,None),
