--- data/ui/widgets.ui
@@ -1,5 +1,0 @@
-<?xml version='1.0' encoding='UTF-8'?>
-<!-- Created with Cambalache 0.16.0 -->
-<interface>
-  <!-- interface-name widgets.ui -->
-</interface>
