+++ docoloco/window/__init__.py
@@ -1,0 +1,3 @@
+from .application import ApplicationWindow
+from .doc_page import DocPage
+from .new_page import NewPage
